Made a working version
Stitched together Robert's file reader with some basic start up stuff. Also fixed the file reader (scope issue with hashMap2) as well as several typos in the Dictionary.txt (namely ":" that improperly delimited). Also turned the file reader into a class with a few functions, namely one to return the Dictionary and another to return the Keys and their synonyms.

To play around with this version, run the program and type into the console. You'll need two lines to elicit a response: the first is the Important Noun, and the second is the W.

I haven't done anything to parse user input with any intelligence, so we still need to hook that up.