Added text parsing
Created a new class: ParseInput. It does as it says. Given a Scanner and an ArrayList of Keys (as supplied by the ReadFile.loadKeys function), this class can prompt the user for an input, and pull out an important set of keys. It will then return those keys as a String array of length 2.