Altered the pipeline, altered negative response, changed buildpath
Build path was changed. In the version most recently pushed, it used Java14 (unbound). Changed it to Java13.

Made it so that a negative response output a secondary key of generic, to prevent it mistakenly picking out a bad secondary.

Defined a pipeline during startup that is passed around; mostly in hopes to make it not run so incredibly slow (I don't think it sped it up). Reflected this change within the method of ParseInputTest.testParseInput().

Finally, added additional characters for quotation marks within the ReadFile.loadDict() method as somehow there are three separate characters within the Dict for quotation marks.