Added Wolfram & Wikipedia API
Added the ability to request XML formatted calls from Wolfram and Wikipedia. The user is required to input a String of the form of "what is/are ?" or "define ?" respectively.

Two stages: first, in ParseInput, the relevant keys are extracted. Second, in StartUp, if an appropriate key is detected, performs an API call